Fix windows CI
test=develop

+++ python/paddle/fluid/tests/unittests/test_feed_data_check_shape_type.py
@@ -88,7 +88,7 @@ class TestFeedData(unittest.TestCase):
             size=[batch_size, 3, 4, 5]).astype(np.float32)
         label_size = [-1, 1]
         feed_label = np.random.randint(
-            low=0, high=self.class_num, size=[batch_size, 1])
+            low=0, high=self.class_num, size=[batch_size, 1]).astype(np.int64)
         self._feed_data_in_executor(in_size, label_size, feed_in_data,
                                     feed_label, use_cuda, use_parallel_executor)
 
@@ -99,7 +99,7 @@ class TestFeedData(unittest.TestCase):
             size=[batch_size, 3, 4, 5]).astype(np.float32)
         label_size = (-1, 1)
         feed_label = np.random.randint(
-            low=0, high=self.class_num, size=[batch_size, 1])
+            low=0, high=self.class_num, size=[batch_size, 1]).astype(np.int64)
         self._feed_data_in_executor(in_size, label_size, feed_in_data,
                                     feed_label, use_cuda, use_parallel_executor)
